+++ libs/platform/src/lib/table/table.component.ts
@@ -214,11 +214,13 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
     @Input()
     enableDragResize = true;
 
-    /** Whether to fix the table header and footer. Default is false.
+    /**
+     * Whether to fix the table header and footer. Default is true.
      * Note that if the table contains freezable columns, the header and
-     * footer will be fixed automatically. */
+     * footer will be fixed automatically, regardless of this input value.
+     * */
     @Input()
-    fixed = false;
+    fixed = true;
 
     /**
      * Table data source.
@@ -1620,10 +1622,20 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
      * @hidden
      * Create table rows rearrange event
      */
-    _emitRowsRearrangeEvent(row: TableRow, previousIndex: number, newIndex: number): void {
+    _emitRowsRearrangeEvent(row: TableRow, dropRow: TableRow, event: FdDropEvent<TableRow>): void {
         const rows = this._tableRows.map(({ value }) => value);
 
-        this.rowsRearrange.emit(new TableRowsRearrangeEvent(row.value, previousIndex, newIndex, rows));
+        this.rowsRearrange.emit(
+            new TableRowsRearrangeEvent(
+                row.value,
+                dropRow.value,
+                event.draggedItemIndex,
+                event.replacedItemIndex,
+                event.insertAt,
+                event.mode,
+                rows
+            )
+        );
     }
 
     /**
@@ -1765,7 +1777,7 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
             }
 
             this._cdr.markForCheck();
-            this._emitRowsRearrangeEvent(dragRow, event.draggedItemIndex, event.replacedItemIndex);
+            this._emitRowsRearrangeEvent(dragRow, dropRow, event);
         }
     }
 
@@ -2037,7 +2049,7 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
         this._subscriptions.add(
             this._tableService.freezeChange.subscribe((event: FreezeChange) => {
                 this.columnFreeze.emit(new TableColumnFreezeEvent(this, event.current, event.previous));
-                this.fixed = !!this._freezableColumns.size || !!this._freezableEndColumns.size;
+                this.fixed = !!this.fixed || !!this._freezableColumns.size || !!this._freezableEndColumns.size;
             })
         );
 
@@ -2318,7 +2330,7 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
     private _setFreezableInfo(): void {
         this._freezableColumns = this._getFreezableColumns();
         this._freezableEndColumns = this._getFreezableEndColumns();
-        this.fixed = !!this._freezableColumns.size || !!this._freezableEndColumns.size;
+        this.fixed = !!this.fixed || !!this._freezableColumns.size || !!this._freezableEndColumns.size;
     }
 
     /** @hidden */